+++ unit_tests/environments/sugar/string.js
@@ -1029,35 +1029,5 @@ test('String', function () {
   equal('{one {1} {2} two}'.assign(5,6), '{one 5 6 two}', 'String#assign | nested braces with strings outside');
 
 
-  // String#namespace (JS equivalent of ActiveSupport::Inflector.safe_constantize)
-
-  equal('foobar'.namespace(), undefined, 'String#namespace | undefined namespace')
-
-  // Global
-  Foo = 'bar';
-
-  equal('Foo'.namespace(), 'bar', 'String#namespace | simple namespace')
-
-  // Global
-  Foo = {
-    Bar: {
-      Moo: {
-        Car: 'wasabi',
-        Fn: function(){},
-        NULL: null
-      }
-    }
-  };
-
-  equal('Foo.Bar.Moo.Car'.namespace(), 'wasabi', 'String#namespace | deep namespace')
-  equal('Foo.Bar.Moo.Fn'.namespace(), Foo.Bar.Moo.Fn, 'String#namespace | function')
-  equal('Foo.Bar.Moo.NULL'.namespace(), null, 'String#namespace | null')
-  equal('Dumb.Bar.Moo'.namespace(), undefined, 'String#namespace | non-existant top')
-  equal('Foo.Dumb.Moo'.namespace(), undefined, 'String#namespace | non-existant middle')
-  equal('Foo.Bar.Dumb'.namespace(), undefined, 'String#namespace | non-existant bottom')
-  equal('Moo'.namespace(Foo.Bar), Foo.Bar.Moo, 'String#namespace | allows starting scope')
-
-
-
 
 });